requirements.txt
Flask
PyJWT
Flask-WTF 